Factor out link state changes to a separate function
We may want to modify the link state on several moments, so don't
implement it over and over and over and over again every time we want to
do so. Also, on lo all we need to do is bring the interface up -- the
kernel will do the rest -- so don't try to set an address.