initial import
Took some parts from the original ipcfg, but nuked most of it. This
really is a fresh start.

We'll probably look at the old code to remember how we did things, but:
- The backend is using libnl in caching mode, which means we won't see
  changes (and we want to see changes). Using it in non-cachine mode
  requires significant changes to the whole system.
- The middle end is based on a broken concept. That needs to be redone,
  too.
- The frontend wasn't bad per se, but our new frontend wouldn't be
  entirely compatible.
- There were some utility functions (hashtable, doubly-linked list, some
  macros) that weren't really ipcfg-specific; so we kept those, but it's not
  much.

Hence, rather than using the same git repo, I created a new one. This
really is different code. Yes, I should probably have given it a
different name, but it's not as if anyone has been using ipcfg yet,
anyway.

No, it doesn't compile yet.